Fixed problems with merging and shifted Megan's models to core
