+++ core/static/css/pages/lesson.scss
@@ -14,7 +14,6 @@
     height: calc(100vh - #{$page-header-min-height + $page-footer-height});
 }
 
-
 #tutorialContent {
     overflow-x: hidden;
     overflow-y: auto;
