--- mysite/templates/python_example_2_r5TgYkn.txt
@@ -1,6 +1,0 @@
-class Code_Template(models.Model):
-	template_name = models.CharField(max_length=30)
-	template_code = models.TextField(max_length=500)
-
-	def __str__(self):
-		return self.template_name
